Trekk ut hjelpefunksjonar for enslig forsørger i ei fil slik at vi lettare kan sjå kva vi gjer med dataa vi får frå backend, og lettare kan dele denne koden med samarbeidspartnarar.
